Move code from extension.ts and import from utils
